backend/requirements.txt
appnope==0.1.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bcrypt==3.1.6
cffi==1.12.2
Click==7.0
decorator==4.4.0
fastapi==0.12.1
h11==0.8.1
httptools==0.0.13
ipdb==0.12
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
more-itertools==7.0.0
parso==0.4.0
passlib==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.9.0
pony==0.7.9
prompt-toolkit==2.0.9
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
pydantic==0.21
Pygments==2.3.1
PyJWT==1.7.1
pytest==4.4.0
python-decouple==3.1
six==1.12.0
starlette==0.11.1
traitlets==4.3.2
uvicorn==0.6.1
uvloop==0.12.2
wcwidth==0.1.7
websockets==7.0
